Improve compatibility with other SPI libraries
One of the things speeding up this library is that it just waits the bare minimum needed to send the next command to the LCD instead of waiting for the SPI controller to flag a finished transmission. While that works fine with the display it appears that not everything is finished after that timespan. Calling endTransaction before the SPI controller completely finished the transmission results in other communication partners on the same bus behaving erratically. This line shouldn't add much delay because it is only executed after the complete transmission finished, but it allows other controllers on the same bus to be properly separated via Transactions.
I am specifically using the touch library by Paul Stoffregen https://github.com/PaulStoffregen/XPT2046_Touchscreen because I want all chips to run on a single SPI bus to save pins. This library won't work without this added line.